Present current storage stats
